ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
markdown==3.5.2
reportlab==4.0.7
python-multipart==0.0.6

# Optional: WeasyPrint for higher quality PDFs (requires GTK3 on Windows)
# weasyprint==60.1
